engine/pyproject.toml
[tool.poetry]
name = "monster-engine"
version = "0.1.0"
description = ""
authors = ["Scott Persinger <scottpersinger@gmail.com>"]
readme = "README.md"
#package-mode = false

[tool.poetry.dependencies]
python = "^3.11,<3.13"
fastapi = "^0.109.2"
sqlalchemy = "^2.0.26"
psycopg2-binary = "^2.9.9"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlmodel = "0.0.22"
pytest = "^8.0.0"
httpx = ">=0.27.0,<0.28.0"
slack-bolt = "^1.18.1"
slack-sdk = "^3.26.2"
redis = "4.6.0"
pandas = "^2.2.0"
cryptography = "^42.0.2"
pyarrow = "^15.0.0"
jira = "^3.6.0"
html2text = "^2020.1.16"
google-api-core = "^2.17.0"
google-api-python-client = "^2.118.0"
google-auth-oauthlib = "^1.2.0"
serpapi = "^0.1.5"
google-search-results = "^2.4.2"
python-dotenv = "^1.0.1"
alembic = "^1.13.1"
flask = "^3.0.2"
flask-dance = "^7.0.1"
simple-salesforce = "^1.12.5"
fastapi-lifespan-manager = "^0.1.4"
boto3 = "^1.34.56"
openpyxl = "^3.1.2"
apscheduler = "^3.10.4"
simplegmail = "^4.1.1"
google-news-feed = "^1.1.0"
markdown2 = "^2.4.13"
bleach = "^6.1.0"
python-nmap = "^0.7.1"
aiofiles = "^23.2.1"
python-multipart = "^0.0.9"
playwright = "^1.43.0"
snowflake-connector-python = {extras = ["pandas"], version = "^3.8.1"}
duckdb = "^0.10.1"
pytextract = "^2.0.1"
langchain-openai = "0.1.9"
langchain-groq = "0.1.9"
async-simple-salesforce = "^1.12.4a1"
langchain = "0.2.5"
langchain-core = "0.2.26"
gitpython = "^3.1.43"
pydub = "^0.25.1"		
multiprocess = "^0.70.16"
scipy = "^1.13.1"
cloudpickle = "^3.0.0"
sounddevice = "^0.4.7"
builtwith = "^1.3.4"
langchain-anthropic = "0.1.22"
langchain-community = "0.2.5"
langsmith = "^0.1.90"
e2b = "0.17.1"
e2b-code-interpreter = "0.0.10"
replicate = "^0.27.0"
docker = "^7.1.0"
pytest-asyncio = "^0.23.7"
ipython = "^8.26.0"
premailer = "^3.10.0"
css-inline = "^0.14.1"
pypdf2 = "^3.0.1"
plotly = "^5.23.0"
matplotlib = "^3.9.2"
rollbar = "^1.0.0"
python-levenshtein = "^0.25.1"
simplejson = "^3.19.3"
hubspot-api-client = "^9.0.0"
setuptools = "^74.0.0"
aioduckdb = "^0.1.0"
pgvector = "^0.2.4"
croniter = "^3.0.3"
pypdf = "^4.3.1"
yt-dlp = "^2024.8.6"
whisper = "^1.1.10"
pytube = "^15.0.0"
firecrawl-py = "^1.2.3"
pymysql = "^1.1.1"
pymssql = "^2.3.1"
fpdf2 = "^2.8.1"
pymupdf = "^1.24.11"
caldav = "^1.3.9"
langchainhub = "^0.1.21"
llama-index = "^0.11.19"
llama-index-vector-stores-postgres = "^0.2.6"
llama-index-readers-notion = "^0.2.0"
pillow = "^11.0.0"
anthropic = "^0.37.1"
ragie = "^1.2.1"
zaproxy = "^0.3.2"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"